Changed specification and name of readBoardFile.
To better represent what the function does, function name was changed in
specification to readBoardsFile, and corrected to this in function placeholder.

To better describe what a board file is, the parameter name was changed to
boardsFile, file extention was changed from .board to .boards to represent that
a boards file can contain several boards.