Changed impl. and spec. of readNumbersFromLine.
As hexadecimal sudoku will use 0-F, and others may use any symbols, better use
'a option as return value. Using int option at the moment for simplicity, but
implementation could be changed to take a list of symbols (or let a higher
function "guess" symbols) as indata or simply return a list of chars or strings.